fix(firmware): Revert change to GptNextKernelEntry made in e3ecb4d9.
We rely on the new cgpt next command which doesn't use this related code
in cgptlib. Since this change breaks existing unit tests but isn't
required we can roll it back. Ideally all this kernel partition code
should be more generic so any partition type can be used but that's a
big change that we don't need right now.