fix(cgpt): Add support for repairing resized disks.
As-is `cgpt repair` when confronted with a disk with a new size will
write the existing table with the old sizes, leaving the disk every bit
as broken as it was before. Now in addition to replacing the secondary
header it will properly recompute the properties that depend on disk
size and update both headers.