feat(cgpt): Generate hybrid MBR partition table.
If a partition has the "Legacy BIOS Bootable" flag set add it to the
MBR partition table. This replaces the need to use an external tool such
as gptfdisk to create hybrid MBRs for pvgrub and ensures that the MBR
remains valid after `cgpt resize` or `cgpt repair`.

The wonkiest part of this code is just computing the CHS values.