cgpt: simplify protective and hybrid MBRs
When generating hybrid MBRs (the default for CoreOS images) the table
will no longer include an extra 0xEE type partition to cover the space
between the MBR and the first partition. Generating a table in this way
causes the partitions to appear out of order in the MBR which may or may
not be a problem. More importantly gdisk warns that two 0xEE partitions
can cause problems with some systems.

When generating a simple protective MBR the table now matches the output
of parted exactly, using bogus values for the CHS fields. On CoreOS to
use this type of MBR run the following command which clears the Legacy
BIOS Bootable flag on the boot partition:

    cgpt add -i 1 -B 0 /dev/sda

Hopefully either or both of these changes will help systems that can
boot when the partition table was generated by parted but failed with
the current cgpt. Reported in https://github.com/coreos/bugs/issues/159